Started implementing type hierarchy.
git-svn-id: http://svn.dsource.org/projects/descent/trunk@1478 6538119e-2221-0410-ba12-9a5a7e5062c7

+++ descent.core/src/descent/core/ITypeHierarchy.java
@@ -145,7 +145,7 @@ IType[] getAllTypes();
  * @see Flags
  * @since 2.0
  */
-int getCachedFlags(IType type);
+long getCachedFlags(IType type);
 
 /**
  * Returns all interfaces resolved to extend the given interface,
